+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * Migrating functions from **ClimMobTools** to **gosset**. ClimMobTools will keep only the functions exclusively related to the 'ClimMob' platform. Other functions are transferred to **gosset** to provide a better environment for data handling, analysis and visualization.
 * Retain function `getDataCM`, `getProjectCM`, `randomise` and `seed_need`
+* Changes in package description
+* Internal functions for pivoting data.frames to avoid dependencies
 
 # ClimMobTools 0.2-8